pa/genetic_algorithm/R2/requirements.txt
aniso8601==3.0.2
certifi==2018.4.16
chardet==3.0.4
click==6.7
docopt==0.6.2
enum34==1.1.6
Flask==1.0.2
Flask-RESTful==0.3.6
haversine==2.1.2
httputils==0.1b0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pipreqs==0.4.9
pymongo==3.7.1
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
urllib3==1.23
uuid==1.30
Werkzeug==0.14.1
yarg==0.1.9
